apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "FastAPI backend"
authors = ["Developer <dev@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
prisma = "^0.11.0"
watchfiles = "^0.21.0"
python-dotenv = "^1.0.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 